Create Arrange-Students
Credit: HackerRank
A classroom has several students, half of whom are boys and half of whom are girls. You need to arrange all of them in a line for the morning assembly such that the following conditions are satisfied:
- The students must be in order of non-decreasing height.
- Two boys or two girls must not be adjacent to each other.

You have been given the heights of the boys in the array b and the heights of the girls in the array g. Find out whether you can arrange them in an order which satisfies the given conditions. Print "YES" if it is possible, or "NO" if it is not.

For example, let's say there are n = 3 boys and n = 3 girls, where the boys' heights are b = [5,3,8] and the girls' heights are g = [2,4,6]. These students can be arranged in the order [g0,b1,g1,b0,g2,b2], which is [2, 3, 4, 5, 6, 8]. Because this is in order of non-decreasing height, and no two boys or two girls are adjacent to each other, this satisfies the conditions. Therefore, the answer is "YES".

Input Format:
- The first line contains an integer, t, denoting the number of test cases.
- The first line of each test case contains an integer, n, denoting the number of boys and girls in the classroom.
- The second line of each test case contains n space separated integers, b1, b2, ...bn, denoting the heights of the boys.
- The third line of each test case contains n space separated integers, g1, g2, ...gn, denoting the heights of the girls.

Constraints:
- 1<=t<=10
-1<=n<=100

Output Format:
Print exactly t lines. In the ith of them, print a single line containing "YES" if it is possible to arrange the students in the ith test case, or "NO" if it is not.